+++ static_precompiler/compilers/libsass.py
@@ -19,9 +19,10 @@ class SCSS(scss.SCSS):
     indented = False
     is_compass_enabled = False
 
-    def __init__(self, sourcemap_enabled=False, load_paths=None, precision=None):
+    def __init__(self, sourcemap_enabled=False, load_paths=None, precision=None, output_style=None):
         self.is_sourcemap_enabled = sourcemap_enabled
         self.precision = precision
+        self.output_style = output_style
         if load_paths is None:
             self.load_paths = []
         elif not isinstance(load_paths, (list, tuple)):
@@ -54,6 +55,8 @@ class SCSS(scss.SCSS):
                     compile_kwargs["include_paths"] = self.load_paths
                 if self.precision:
                     compile_kwargs["precision"] = self.precision
+                if self.output_style:
+                    compile_kwargs["output_style"] = self.output_style
                 compiled = sass.compile(filename=full_source_path, **compile_kwargs)
         except sass.CompileError as e:
             raise exceptions.StaticCompilationError(encoding.force_str(e))