+++ CHANGES.rst
@@ -6,14 +6,15 @@ Dev
 ===
 
 - Fix Stylus compiler to actually enable support for detecting changes in imported files
-- Add `precision` option to SASS / SCSS / LibSass compilers. Set it to 8 or more if you compile Bootstrap.
+- Add ``precision`` option to SASS / SCSS / LibSass compilers. Set it to 8 or more if you compile Bootstrap.
+- Add ``output_style`` option to SASS / SCSS / LibSass compilers.
 
 
 1.2
 ===
 
 - Add LiveScript compiler
-- Add support for `--global-var` option in LESS compiler
+- Add support for ``--global-var`` option in LESS compiler
 - Add SCSS / SASS compiler based on Libsass